VULCAN-19 Move TriConsumer to public functions package
