+++ modules/apps/foundation/vulcan/vulcan-jaxrs-writer-json/src/main/java/com/liferay/vulcan/jaxrs/writer/json/internal/WriterHelper.java
@@ -19,6 +19,7 @@ import static org.osgi.service.component.annotations.ReferencePolicyOption.GREED
 
 import com.liferay.vulcan.error.VulcanDeveloperError;
 import com.liferay.vulcan.filter.QueryParamFilterType;
+import com.liferay.vulcan.function.TriConsumer;
 import com.liferay.vulcan.list.FunctionalList;
 import com.liferay.vulcan.resource.Resource;
 import com.liferay.vulcan.response.control.Embedded;
